fix(topology): avoid infinite loop on PCI bridge with subordinate bus 0xff

+++ src/application/topology/pci.cpp
@@ -297,9 +297,12 @@ void TopoSystemPci::Load() {
     pcis.emplace(node->BusId().packed, node);
 
     if (headerType == PCI_HEADER_TYPE_BRIDGE) {
-      uint8_t secondary = pci_read_byte(dev, PCI_SECONDARY_BUS);
-      uint8_t subordinate = pci_read_byte(dev, PCI_SUBORDINATE_BUS);
-      for (uint8_t i = secondary; i <= subordinate; i++) {
+      // Use a wider counter than the bus numbers themselves: a bridge may report
+      // a subordinate bus of 0xff (e.g. empty downstream ports on some NICs), and
+      // a uint8_t loop variable would wrap from 0xff back to 0 and spin forever.
+      uint32_t secondary = pci_read_byte(dev, PCI_SECONDARY_BUS);
+      uint32_t subordinate = pci_read_byte(dev, PCI_SUBORDINATE_BUS);
+      for (uint32_t i = secondary; i <= subordinate; i++) {
         uint64_t dspBusId = PciBusId(dev->domain, i, 0, 0).packed;
         if (dsp2dev.find(dspBusId) != dsp2dev.end()) {
           struct pci_dev* lastDev = dsp2dev[dspBusId];
